Standard memoization syntax for fetch_bag; better docs; moving to presenters dir.
